Add note for the Content type to sync
In Intune admin portal when an Android Email profile is created for Knox the settings available are "Enable" and "Not Configured" but the  actions defined on the code is Disable. 

The note is with the intention to clarify the action on the background and prevent confusions for customers and support team.